Modified COM interfaces (IUnknown and others) to use UINT32 instead of ULONG. On 32-bit ULONG is 32 bits long in C++; this gets translated to UInt32 in .NET. However, on 64-bit Linux machines ULONG is 64 bits long, therefore we have to use UINT32 in all COM interfaces that are used in managed and unmanaged code.
Reviewed by: Tom Hindle


svn-revision: 456

+++ COM/ManagedComBridge/ManagedComBridge.cpp
@@ -140,7 +140,7 @@ STDMETHODIMP ManagedComBridge::QueryInterface(REFIID iid, void ** ppv)
 /**
  * Reference counting not implemented for this factory class. 
  */
-ULONG ManagedComBridge::AddRef(void)
+UINT32 ManagedComBridge::AddRef(void)
 {
 	return 0;	
 }
@@ -148,7 +148,7 @@ ULONG ManagedComBridge::AddRef(void)
 /**
  * Reference counting not implemented for this factory class. 
  */
-ULONG ManagedComBridge::Release(void)
+UINT32 ManagedComBridge::Release(void)
 {
 	return 0;	
 }
